Samples/VPC/Python/Vanilla/requirements.txt
aws-cdk-lib==2.26.0
constructs>=10.0.0,<11.0.0
